requirements.txt
telethon>=1.24.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
psutil>=5.8.0
asyncio
aiofiles
aiohttp
cryptography
requests
schedule
sqlalchemy
